aquarius: ♒ initial commit
`aquarius` is a command-line tool & library for http/2 load-testing.

this initial commit introduces the basic initial project structure, and
a proof-of-concept implementation. to start a local http/2 server, and
run a load-test against it, run the following commands in two terminal
windows:

```
just run-test-server
just run-load-test
```

 ### 📂 project structure

source code for `aquarius` lives in `src/`. integration tests can be
found in `tests/`.

the `aquarius` crate may also be used as a library for use in other
load-testing applications, or in integration tests. the `crates/`
directory contains some other small tools to facilitate testing: an
http/2 server, a RAII-bounded `tracing-subscriber` that is compatible
with `libtest`'s output capturing, and a timeout fuse to indicate
failure when a test may have deadlocked. these libraries are development
dependencies of `aquarius`.

a nix flake providing a development shell is also included, alongside a
`justfile` that can help run common commands.

```
; eza --tree --level=2 --git-ignore
.
├── Cargo.lock
├── Cargo.toml
├── crates
│  ├── test-server
│  ├── test-subscriber
│  └── test-timeout
├── doc
│  └── example.png
├── flake.lock
├── flake.nix
├── justfile
├── README.md
├── rust-toolchain.toml
├── src
│  ├── charts
│  ├── charts.rs
│  ├── cli.rs
│  ├── lib.rs
│  ├── main.rs
│  ├── summary
│  ├── summary.rs
│  ├── syndicate
│  ├── syndicate.rs
│  ├── worker
│  └── worker.rs
└── tests
   ├── can_send_many_requests.rs
   └── can_send_one_request.rs
```